+++ src/test/java/filehandler/algorithm/SplitAlgorithmsTest.java
@@ -10,7 +10,7 @@ import static org.junit.Assert.*;
  * Created by mzeus on 7/3/16.
  */
 public class SplitAlgorithmsTest {
-    SplitAlgorithms<Integer> algorithm = new SplitAlgorithms();
+    SplitAlgorithms<Byte> algorithm = new SplitAlgorithms();
 
     @Before
     public void warmUp() {
@@ -20,22 +20,22 @@ public class SplitAlgorithmsTest {
 
     @Test
     public void decrypt() throws Exception {
-        assertNotEquals(algorithm.decrypt(20, 0, 0), new Integer(20));
-        assertNotEquals(algorithm.decrypt(20, 0, 1), new Integer(20));
-        assertNotEquals(algorithm.decrypt(20, 0, 1), algorithm.decrypt(20, 0, 0));
+        assertNotEquals(algorithm.decrypt((byte) 20, (byte) 0, 0), new Integer(20));
+        assertNotEquals(algorithm.decrypt((byte) 20, (byte) 0, 1), new Integer(20));
+        assertNotEquals(algorithm.decrypt((byte) 20, (byte) 0, 1), algorithm.decrypt((byte) 20, (byte) 0, 0));
     }
 
     @Test
     public void encryptEqualsDecrypt() {
-        assertEquals(algorithm.encrypt(algorithm.decrypt(20, 0, 0), 0, 0).byteValue(), new Integer(20).byteValue());
+        assertEquals(algorithm.encrypt(algorithm.decrypt((byte) 20, (byte) 0, 0), (byte) 0, 0).byteValue(), new Integer(20).byteValue());
     }
 
 
     @Test
     public void encrypt() throws Exception {
-        assertNotEquals(algorithm.encrypt(20, 0, 0), new Integer(20));
-        assertNotEquals(algorithm.encrypt(20, 0, 1), new Integer(20));
-        assertNotEquals(algorithm.encrypt(20, 0, 1), algorithm.encrypt(20, 0, 0));
+        assertNotEquals(algorithm.encrypt((byte) 20, (byte) 0, 0), new Integer(20));
+        assertNotEquals(algorithm.encrypt((byte) 20, (byte) 0, 1), new Integer(20));
+        assertNotEquals(algorithm.encrypt((byte) 20, (byte) 0, 1), algorithm.encrypt((byte) 20, (byte) 0, 0));
     }
