+++ src/test/java/filehandler/algorithm/cipheralgorithm/MultiplicationAlgorithmTest.java
@@ -15,25 +15,25 @@ public class MultiplicationAlgorithmTest {
     @Test
     public void encryptEqualsDecrypt() {
         algorithm.generateEncryptKeys();
-        assertEquals(algorithm.encrypt(algorithm.decrypt(20, 0, 0), 0, 0).byteValue(), new Integer(20).byteValue());
+        assertEquals(algorithm.encrypt(algorithm.decrypt((byte) 20, (byte) 0, 0), (byte) 0, 0).byteValue(), new Integer(20).byteValue());
     }
 
     @Test
     public void decrypt() throws Exception {
         algorithm.generateEncryptKeys();
-        assertNotEquals(algorithm.decrypt(20, 20, 0).byteValue(), new Integer(20).byteValue());
+        assertNotEquals(algorithm.decrypt((byte) 20, (byte) 20, 0).byteValue(), new Integer(20).byteValue());
     }
 
     @Test
     public void encrypt() throws Exception {
         algorithm.generateEncryptKeys();
-        assertNotEquals(algorithm.encrypt(20, 20, 0).byteValue(), new Integer(20).byteValue());
+        assertNotEquals(algorithm.encrypt((byte) 20, (byte) 20, 0).byteValue(), new Integer(20).byteValue());
     }
 
     @Test
     public void checkIfKeyIsValid() throws Exception {
-        assertFalse(algorithm.checkIfKeyIsValid(20));
-        assertTrue(algorithm.checkIfKeyIsValid(1));
+        assertFalse(algorithm.checkIfKeyIsValid((byte) 20));
+        assertTrue(algorithm.checkIfKeyIsValid((byte) 1));
 
     }
